Move Pool Retire components to own folder
- Preparation for future Retirements from Listings
